Add getResponse() to Som and update relevant classes for GUI integration
The Som chatbot used to be CLI based, which used System.out.println(), which does not enable the GUI to display responses
in chat bubbles.

Let's add a new method getResponse(String input) to return bot responses as strings by:
- Updating all relevant Command class and subclasses
- Updating the Som class to support getResponse()
- Updating the Ui class to return strings instead of printing out the results

This enables the GUI to display responses in chat bubbles.